+++ tools/pybullet_linear_tire_example.py
@@ -53,7 +53,7 @@ def look_at(client, position=(0, 0, 0), top_down=True):
         )
 
 
-def run(base, client, vehicle, plane_body_id, sliders, n_steps=1e6):
+def run(renderer, client, vehicle, plane_body_id, sliders, n_steps=1e6):
     prev_friction_sum = None
 
     controller_state = ActuatorDynamicControllerState()
@@ -168,7 +168,7 @@ def run(base, client, vehicle, plane_body_id, sliders, n_steps=1e6):
 
         client.stepSimulation()
         vehicle.sync_to_panda3d()
-        showbase.taskMgr.step()
+        renderer.step()
 
         frictions_ = frictions(sliders)
 
@@ -210,8 +210,8 @@ if __name__ == "__main__":
     # https://turtlemonvh.github.io/python-multiprocessing-and-corefoundation-libraries.html
     # mp.set_start_method('spawn', force=True)
 
-    # showbase = ShowBase(windowType="onscreen")
-    showbase = mock.MagicMock()
+    # renderer = Renderer("linear_tier_example")
+    renderer = mock.MagicMock()
 
     client = bc.BulletClient(pybullet.GUI)
     # client = BulletClient(pybullet.GUI)
@@ -256,7 +256,7 @@ if __name__ == "__main__":
             vehicle = Vehicle(
                 "hello",
                 pose=pose,
-                showbase=showbase,
+                renderer=renderer,
                 chassis=AckermannChassis(
                     pose=pose,
                     bullet_client=client,
@@ -265,7 +265,7 @@ if __name__ == "__main__":
             )
 
             run(
-                showbase,
+                renderer,
                 client,
                 vehicle,
                 plane_body_id,