+++ tools/pybullet_trajectory_example.py
@@ -35,7 +35,7 @@ def look_at(client, position=(0, 0, 0), top_down=True):
         )
 
 
-def run(base, client, vehicle, plane_body_id, sliders, n_steps=1e6):
+def run(renderer, client, vehicle, plane_body_id, sliders, n_steps=1e6):
     prev_friction_sum = None
 
     controller_state = TrajectoryTrackingControllerState()
@@ -90,7 +90,7 @@ def run(base, client, vehicle, plane_body_id, sliders, n_steps=1e6):
 
         client.stepSimulation()
         vehicle.sync_chassis()
-        showbase.taskMgr.step()
+        renderer.step()
 
         frictions_ = frictions(sliders)
 
@@ -134,8 +134,8 @@ if __name__ == "__main__":
     # https://turtlemonvh.github.io/python-multiprocessing-and-corefoundation-libraries.html
     # mp.set_start_method('spawn', force=True)
 
-    # showbase = ShowBase(windowType="onscreen")
-    showbase = mock.MagicMock()
+    # renderer = Renderer("trajsectory_example")
+    renderer = mock.MagicMock()
 
     client = bc.BulletClient(pybullet.GUI)
     # client = BulletClient(pybullet.GUI)
@@ -177,7 +177,7 @@ if __name__ == "__main__":
             vehicle = Vehicle(
                 id="vehicle",
                 pose=pose,
-                showbase=showbase,
+                renderer=renderer,
                 chassis=AckermannChassis(
                     pose=pose,
                     bullet_client=client,
@@ -185,7 +185,7 @@ if __name__ == "__main__":
             )
 
             run(
-                showbase,
+                renderer,
                 client,
                 vehicle,
                 plane_body_id,